add lint prettier in Ci (#3262)
* add lint prettier in Ci

* .circleci use version =2

* remove jest in package

* update package.json

* add enzyme

* fix appveyor error

* use check-prettier script
