+++ .circleci/config.yml
@@ -6,6 +6,7 @@ jobs:
     steps:
       - checkout
       - run: npm install
+      - run: npm run lint
       - run: npm run build
   test:
     docker:
@@ -13,9 +14,9 @@ jobs:
     steps:
       - checkout
       - run: npm install
-      - run: 
-          command : npm run test:all
-          no_output_timeout : 30m
+      - run:
+          command: npm run test:all
+          no_output_timeout: 30m
 workflows:
   version: 2
   build_and_test:
