+++ appveyor.yml
@@ -1,10 +1,10 @@
 # Test against the latest version of this Node.js version
 environment:
-  nodejs_version: "8"
+  nodejs_version: '10'
 
 # this is how to allow failing jobs in the matrix
 matrix:
-  fast_finish: true     # set this flag to immediately finish build once one of the jobs fails.
+  fast_finish: true # set this flag to immediately finish build once one of the jobs fails.
 
 # Install scripts. (runs after repo cloning)
 install:
