+++ package.json
@@ -11,7 +11,8 @@
     "site": "npm run presite && cross-env APP_TYPE=site npm run build && firebase deploy && npm run docker:push",
     "analyze": "cross-env ANALYZE=1 umi build",
     "lint:style": "stylelint \"src/**/*.less\" --syntax less",
-    "lint": "eslint --ext .js src mock tests && npm run lint:style",
+    "lint:prettier": "check-prettier lint",
+    "lint": "eslint --ext .js src mock tests && npm run lint:style && npm run lint:prettier",
     "lint:fix": "eslint --fix --ext .js src mock tests && npm run lint:style",
     "lint-staged": "lint-staged",
     "lint-staged:js": "eslint --ext .js",
@@ -30,11 +31,12 @@
     "docker:push": "npm run docker-hub:build && npm run docker:tag && docker push chenshuai2144/ant-design-pro"
   },
   "dependencies": {
-    "@babel/runtime": "^7.1.5",
-    "antd": "^3.10.9",
-    "bizcharts": "^3.4.0",
-    "bizcharts-plugin-slider": "^2.1.1-beta.1",
     "@antv/data-set": "^0.10.0",
+    "@babel/runtime": "^7.2.0",
+    "antd": "^3.11.6",
+    "bizcharts": "^3.4.2",
+    "bizcharts-plugin-slider": "^2.1.1-beta.1",
+    "check-prettier": "^1.0.1",
     "classnames": "^2.2.6",
     "dva": "^2.4.0",
     "enquire-js": "^0.2.1",
@@ -67,7 +69,7 @@
     "babel-eslint": "^10.0.1",
     "cross-env": "^5.1.1",
     "cross-port-killer": "^1.0.1",
-    "enzyme": "^3.7.0",
+    "enzyme": "3.7.0",
     "eslint": "^5.4.0",
     "eslint-config-airbnb": "^17.0.0",
     "eslint-config-prettier": "^3.0.1",
@@ -85,16 +87,16 @@
     "mockjs": "^1.0.1-beta3",
     "prettier": "1.15.2",
     "pro-download": "^1.0.1",
+    "slash2": "^2.0.0",
     "stylelint": "^9.8.0",
     "stylelint-config-prettier": "^4.0.0",
     "stylelint-config-standard": "^18.0.0",
     "tslint": "^5.10.0",
     "tslint-config-prettier": "^1.10.0",
     "tslint-react": "^3.6.0",
-    "umi": "^2.2.7",
+    "umi": "^2.3.1",
     "umi-plugin-ga": "^1.1.3",
-    "umi-plugin-react": "^1.2.0",
-    "slash2": "^2.0.0"
+    "umi-plugin-react": "^1.2.0"
   },
   "optionalDependencies": {
     "puppeteer": "^1.10.0"
@@ -114,6 +116,13 @@
     "> 1%",
     "last 2 versions",
     "not ie <= 10"
+  ],
+  "checkFiles": [
+    "src/**/*.js*",
+    "src/**/*.ts*",
+    "src/**/*.less",
+    "config/**/*.js*",
+    "scripts/**/*.js"
   ],
   "husky": {
     "hooks": {
