+++ scripts/prettier.js
@@ -9,24 +9,12 @@
 const glob = require('glob');
 const prettier = require('prettier');
 const fs = require('fs');
+const getPrettierFiles = require('./getPrettierFiles');
 const prettierConfigPath = require.resolve('../.prettierrc');
 
 let didError = false;
 
-let files = [];
-const jsFiles = glob.sync('src/**/*.js*', { ignore: ['**/node_modules/**', 'build/**'] });
-const tsFiles = glob.sync('src/**/*.ts*', { ignore: ['**/node_modules/**', 'build/**'] });
-const configFiles = glob.sync('config/**/*.js*', { ignore: ['**/node_modules/**', 'build/**'] });
-const scriptFiles = glob.sync('scripts/**/*.js');
-const lessFiles = glob.sync('src/**/*.less*', { ignore: ['**/node_modules/**', 'build/**'] });
-files = files.concat(jsFiles);
-files = files.concat(tsFiles);
-files = files.concat(configFiles);
-files = files.concat(scriptFiles);
-files = files.concat(lessFiles);
-if (!files.length) {
-  return;
-}
+const files = getPrettierFiles();
 
 files.forEach(file => {
   const options = prettier.resolveConfig.sync(file, {
